change: return CoroutineScope-based Loadable-Flow-creator functions return type to MutableStateFlow
