+++ buildSrc/src/main/java/Versions.kt
@@ -16,8 +16,8 @@ object Versions {
     val java = JavaVersion.VERSION_11
 
     object Loadable {
-        const val CODE = 18
-        const val NAME = "1.6.5"
+        const val CODE = 19
+        const val NAME = "1.6.6"
         const val SDK_COMPILE = 33
         const val SDK_MIN = 21
         const val SDK_TARGET = SDK_COMPILE
